v2.1 Final?

+++ modules/accounts.lua
@@ -1,5 +1,5 @@
 local db = require('modules.db')
-local Framework = require 'server.newFramework'
+local Framework = require('modules.framework')
 
 local account = {}
 
@@ -27,10 +27,21 @@ CreateThread(function()
         return StopResource(resourceName)
     end
 
-    pcall(MySQL.query.await, 'DELETE FROM bank_transactions WHERE date < (NOW() - INTERVAL 6 MONTH)')
+    local legacyCheck = db.getLegacyAccounts()
+    if legacyCheck then
+        db.migrateData(legacyCheck)
+    else
+        local newSchema = db.checkDatabaseSchema()
+        if not newSchema then
+            db.createDatabaseSchema()
+        end
+    end
+
+    if Config.purgeTransactions then
+        pcall(MySQL.query.await, 'DELETE FROM bank_transactions WHERE date < (NOW() - INTERVAL 6 MONTH)')
+    end
 
     local bankData = db.selectBankAccounts()
-
     for _, v in pairs(bankData) do
         local job = v.id
         account[job] = { --  cachedAccounts[#cachedAccounts+1]
@@ -45,19 +56,17 @@ CreateThread(function()
     end
 end)
 
-
-function account.addPlayerAccount(source)
-    if not source then return print('not source')end
+function account.AddPlayer(source)
+    if not source then return end
 
     source = tonumber(source)
-    print('source', source)
     local id = Framework.getCharId(source)
-
-    if not id then return print('not id', id) end
+    if not id then return print(locale("invalid_account", id)) end
 
     account[source] = {
         id = id,
         isPlayer = true,
+        source = source,
         type = locale("personal"),
         name = Framework.getCharName(source),
         frozen = false,
@@ -65,10 +74,9 @@ function account.addPlayerAccount(source)
         amount = 0,
         transactions = db.getTransactions(id) or {}
     }
-    print('account[source]', json.encode(account[source]))
 end
 
-function account.addBankAccount(id, cid)
+function account.AddBank(id, cid)
     if not id or account[id] then return end
 
     account[id] = {
@@ -81,9 +89,10 @@ function account.addBankAccount(id, cid)
         transactions = db.getTransactions(id) or {},
         creator = cid
     }
+    return account[id]
 end
 
-function account.changeAccountName(id, newId)
+function account.changeName(id, newId)
     if not id or not account[id] then return end
     if not account[id] or account[newId] then return end
 
@@ -93,24 +102,21 @@ function account.changeAccountName(id, newId)
     account[id] = nil
 end
 
-function account.nukeBankAccount(id)
+function account.nuke(id)
     if not id or not account[id] then return end
     account[id] = nil
 end
 
 function account.addMoney(bank, amount, comment)
     local left = account(bank)
-
-    if not left then print("NO GOOD") return end
+    if not left then print(locale("invalid_account", bank)) return end
 
     if left.isPlayer then
-        local success = Framework.addMoney(bank, amount, 'bank', comment)
-
+        local success = Framework.addMoney(left.source, amount, 'bank', comment)
         if not success then return false end
     else
         left.amount += amount
         db.setBankBalance(bank, left.amount)
-
         return {amount = left.amount, added = amount}
     end
 
@@ -127,17 +133,12 @@ end
 
 function account.removeMoney(bank, amount, comment)
     local left = account(bank)
-    print(json.encode(left))
-    if not left then return print("aint no way lil bro")end
-
+    if not left then return print(locale("invalid_account", bank)) end
 
     if left.isPlayer then
-        print('we here aint no way 1?')
-        local success = Framework.removeMoney(bank, amount, 'bank', comment)
-        print(bank, amount, tostring(success))
+        local success = Framework.removeMoney(left.source, amount, 'bank', comment)
         if not success then return false end
     else
-        print('we here aint no way wwww?')
         if left.amount < amount then return end
 
         left.amount -= amount
@@ -151,12 +152,9 @@ end
 
 function account.removeCash(bank, amount, comment)
     local left = account[bank]
-
     if not left or not left.isPlayer then return end
 
-
-    return Framework.removeMoney(bank, amount, 'cash', comment)
+    return Framework.removeMoney(left.source, amount, 'cash', comment)
 end
-
 
 return account
